Add casts to FOR macro to eliminate mixed signage comparison warnings

+++ tweetnacl.c
@@ -1,5 +1,5 @@
 #include "tweetnacl.h"
-#define FOR(i,n) for (i = 0;i < n;++i)
+#define FOR(i,n) for (i = 0;(int)i < (int)n;++i)
 #define sv static void
 
 typedef unsigned char u8;
